pyproject.toml
[tool.poetry]
name = "neotransact"
version = "0.1.0"
description = ""
authors = ["Mohammed Haydar <mohmedhaydar@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
faker = "^33.0.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
